backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
insightface==0.7.3
onnxruntime==1.19.2
numpy==1.26.4
opencv-python-headless==4.10.0.84
pillow==10.4.0
supabase==2.24.0
python-dotenv==1.0.0

